fix(plugins): Reorganize plugin system and exports
Reorganize the plugin directory by splitting the main index file into
smaller, focused modules like tokens, viz-list, and built-in-plugin.
This change resolves circular dependencies and improves the clarity
of service injections and component exports across the codebase.
