--- dashboard/src/components/plugins/index.ts
@@ -1,10 +1,0 @@
-export * from './viz-manager';
-export * from './plugin-context';
-export * from './plugin-data-migrator';
-export * from './hooks';
-export * from './color-manager';
-export * from './operation-manager';
-export { onVizRendered, notifyVizRendered } from './viz-components/viz-instance-api';
-export { ServiceLocator, Token } from './service/service-locator';
-export { useServiceLocator } from './service/service-locator/use-service-locator';
-export type { IServiceLocator, IDisposable } from './service/service-locator';
