+++ dashboard/src/components/plugins/viz-components/merico-panel-groups/editor/viz-merico-panel-groups-editor.tsx
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { useEffect, useMemo } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import { useStorageData } from '~/components/plugins';
+import { useStorageData } from '~/components/plugins/hooks';
 import { VizConfigProps } from '~/types/plugin';
 import { VizConfigBanner } from '../../../editor-components';
 import { getDefaultConfig, VizMericoPanelGroupsConf } from '../type';
