+++ dashboard/src/dashboard-editor/ui/settings/content/edit-panel/viz-conf.tsx
@@ -4,7 +4,8 @@ import { observer } from 'mobx-react-lite';
 import React, { createElement, useContext } from 'react';
 import { PluginVizConfigComponent } from '~/components/panel/plugin-adaptor';
 import { useConfigVizInstanceService } from '~/components/panel/use-config-viz-instance-service';
-import { IPanelInfo, IVizManager, PluginContext } from '~/components/plugins';
+import { IPanelInfo, IVizManager } from '~/components/plugins/viz-manager';
+import { PluginContext } from '~/components/plugins/plugin-context';
 import { ServiceLocatorProvider } from '~/components/plugins/service/service-locator/use-service-locator';
 import { useEditPanelContext } from '~/contexts';
 import { ErrorBoundary } from '~/utils';
