+++ dashboard/src/interactions/operation/operations/index.ts
@@ -1,5 +1,5 @@
 import { OpenLink } from '~/interactions/operation/operations/open-link';
-import { IDashboardOperationSchema } from '~/types/plugin';
+import type { IDashboardOperationSchema } from '~/types/plugin';
 import { ConsoleLog } from './console-log';
 import { OpenView } from './open-view';
 import { SetFilterValues } from './set-filter-values';
