+++ dashboard/src/interactions/operation/operations/open-view.tsx
@@ -1,7 +1,7 @@
 import { Select } from '@mantine/core';
 import { observer } from 'mobx-react-lite';
 import { useTranslation } from 'react-i18next';
-import { useStorageData } from '~/components/plugins';
+import { useStorageData } from '~/components/plugins/hooks';
 import { useEditContentModelContext } from '~/contexts';
 import { IDashboardOperation, IDashboardOperationSchema, IOperationConfigProps } from '~/types/plugin';
 import { getSelectChangeHandler } from '~/utils/mantine';
