+++ dashboard/src/interactions/trigger/constant.ts
@@ -1,0 +1,1 @@
+export const TRIGGERS_KEY = '__TRIGGERS';
