+++ dashboard/src/interactions/trigger/index.ts
@@ -1,1 +1,2 @@
 export * from './trigger-manager-impl';
+export { TRIGGERS_KEY } from './constant';
