+++ dashboard/src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './dashboard';
 export * from './filter';
-export * from './plugin';
 export type { AnyObject, Ready } from './utils';
