+++ dashboard/src/types/plugin/index.ts
@@ -1,6 +1,6 @@
 import { EventEmitter2 } from 'eventemitter2';
 import React from 'react';
-import { IVizManager } from '~/components/plugins';
+import { IVizManager } from '~/components/plugins/viz-manager';
 import { PanelModelInstance } from '~/dashboard-editor/model/panels';
 import { AnyObject } from '~/types';
 import { ITemplateVariable } from '~/utils';
